refactor: Refactored LinkNet
