+++ doctr/models/detection/linknet/__init__.py
@@ -1,0 +1,6 @@
+from .base import *
+
+from doctr.file_utils import is_tf_available, is_torch_available
+
+if is_tf_available():
+    from .tensorflow import *
